requirements.txt
fastapi
uvicorn
httpx
azure-cognitiveservices-speech
openai
python-dotenv
pydantic
diskcache
jinja2
groq
python-dotenv
requests
bs4
python-multipart
fastapi
uvicorn
google-cloud-speech
google-cloud-texttospeech
pydub
aiofiles